Trying and failing to make the tile popping only go when I want to
tile popping should happen when a bigger number is made or when a new tile is added as part of the random generation. I could also have the popping happen when a piece moves somewhere else, and with the way I've done it, I may have to. The issue is that sometimes it doesn't pop in when a new piece is added, because the logic behind it is completely screwed up